moved all inline styles into seperate file
