+++ src/App.tsx
@@ -4,7 +4,7 @@ import { theme } from "./theme"
 import { BrowserRouter, Route, Routes } from "react-router-dom"
 import Signup from "pages/Signup"
 import Signin from "pages/Signin"
-import HomePage from "pages/HomePage"
+import HomePage from "pages/MainScreen"
 import Redirecting from "pages/Redirecting"
 
 const App: FC = () => {
