+++ src/pages/MainScreen/style.tsx
@@ -1,6 +1,7 @@
 import { styled } from "@mui/material/styles"
 import { AppBarProps as MuiAppBarProps } from "@mui/material/AppBar/AppBar"
 import MuiAppBar from "@mui/material/AppBar"
+import { SxProps, Theme } from "@mui/material"
 
 export const drawerWidth = 240
 
@@ -55,3 +56,25 @@ export const DrawerHeader = styled("div")(({ theme }) => ({
 	...theme.mixins.toolbar,
 	justifyContent: "flex-end",
 }))
+
+export const ListItemStyle: SxProps<Theme> = {
+	paddingLeft: "1rem",
+	"& *": {
+		color: "#fff",
+	},
+}
+
+
+export const textWhite: SxProps<Theme> = {
+	color: "#fff",
+}
+
+export const DrawerStyles: SxProps<Theme> = (theme) => ({
+	width: drawerWidth,
+	flexShrink: 0,
+	"& .MuiDrawer-paper": {
+		width: drawerWidth,
+		boxSizing: "border-box",
+		backgroundColor: theme.palette.secondary.main,
+	}
+})
